requirements.txt
flask
flask-cors
scikit-learn
nltk
numpy
google-cloud-aiplatform
google-cloud-firestore
python-dotenv
fpdf
python-docx
gunicorn
flask-sqlalchemy
psycopg2-binary
google-generativeai
